Add const and move dataset generation to other file as per PR
